Added a optional location to the `Block` Objective
Adding the argument
`loc:<location>`
will make the objective to check for a specific location the block is placed/broken in

Adding the argument
`ignorecancel`
will make the objective ignore the cancellation of a break/build event
(e.g. if a block is trying to be broken, but a plugin - let it be worldguard - is preventing the block to be broken, the objective will not count that block. If the argument `ignorecancel` is added, the objective will count that block anyways.